Use @ensdomains/ens-avatar for resolving avatars, remove image fetching endpoint
