+++ package.json
@@ -33,6 +33,7 @@
     "contracts:compile:abi": "typechain --target ethers-v5 \"./modules/contracts/ethers/**/*.json\""
   },
   "dependencies": {
+    "@ensdomains/ens-avatar": "^0.1.6",
     "@ethersproject/experimental": "^5.5.0",
     "@makerdao/dai-ui-icons": "^0.0.39",
     "@makerdao/dai-ui-icons-branding": "^0.0.39",
